+++ projects/ccd-case-ui-toolkit/src/lib/shared/components/palette/case-flag/components/case-flag-table/case-flag-table.component.spec.ts
@@ -1,12 +1,17 @@
 import { CUSTOM_ELEMENTS_SCHEMA } from '@angular/core';
 import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
+import { RpxLanguage, RpxTranslationService } from 'rpx-xui-translation';
+import { BehaviorSubject } from 'rxjs';
 import { MockRpxTranslatePipe } from '../../../../../test/mock-rpx-translate.pipe';
+import { FlagDetail } from '../../domain';
 import { CaseFlagStatus } from '../../enums';
+import { FlagFieldDisplayPipe } from '../../pipes';
 import { CaseFlagTableComponent } from './case-flag-table.component';
 
 describe('CaseFlagTableComponent', () => {
   let component: CaseFlagTableComponent;
   let fixture: ComponentFixture<CaseFlagTableComponent>;
+  let mockRpxTranslationService: any;
   const flagData = {
     flags: {
       partyName: 'John Smith',
@@ -26,11 +31,15 @@ describe('CaseFlagTableComponent', () => {
         status: CaseFlagStatus.ACTIVE
       },
       {
-        name: 'Sign language',
+        name: 'Sign Language Interpreter',
+        name_cy: 'Dehonglydd Iaith Arwyddion',
         subTypeValue: 'British Sign Language (BSL)',
-        subTypeKey: '',
-        otherDescription: '',
-        flagComment: '',
+        subTypeValue_cy: 'Iaith Arwyddion Prydain (BSL)',
+        subTypeKey: 'bfi',
+        otherDescription: 'English description for Other',
+        otherDescription_cy: 'Disgrifiad Cymraeg ar gyfer Arall',
+        flagComment: 'An English comment',
+        flagComment_cy: 'Sylw Cymreig',
         flagUpdateComment: 'Flag update comment for second flag',
         dateTimeModified: new Date('2021-09-09 00:00:00'),
         dateTimeCreated: new Date('2021-09-09 00:00:00'),
@@ -38,16 +47,31 @@ describe('CaseFlagTableComponent', () => {
         hearingRelevant: false,
         flagCode: '',
         status: CaseFlagStatus.ACTIVE
-      }]
+      }] as FlagDetail[]
     },
     pathToFlagsFormGroup: '',
     caseField: null
   };
 
   beforeEach(waitForAsync(() => {
+    const source = new BehaviorSubject<RpxLanguage>('en');
+    let currentLanguage: RpxLanguage = 'en';
+    mockRpxTranslationService = {
+      language$: source.asObservable(),
+      set language(lang: RpxLanguage) {
+        currentLanguage = lang;
+        source.next(lang);
+      },
+      get language() {
+        return currentLanguage;
+      }
+    };
     TestBed.configureTestingModule({
-      schemas: [CUSTOM_ELEMENTS_SCHEMA],
-      declarations: [CaseFlagTableComponent, MockRpxTranslatePipe]
+      schemas: [ CUSTOM_ELEMENTS_SCHEMA ],
+      declarations: [ CaseFlagTableComponent, MockRpxTranslatePipe, FlagFieldDisplayPipe ],
+      providers: [
+        { provide: RpxTranslationService, useValue: mockRpxTranslationService }
+      ]
     })
     .compileComponents();
   }));
@@ -56,10 +80,12 @@ describe('CaseFlagTableComponent', () => {
     fixture = TestBed.createComponent(CaseFlagTableComponent);
     component = fixture.componentInstance;
     component.caseFlagsExternalUser = false;
+    // Set default translation language to English
+    mockRpxTranslationService.language = 'en';
     fixture.detectChanges();
   });
 
-  it('should create', () => {
+  it('should create component', () => {
     expect(component).toBeTruthy();
   });
 
@@ -131,5 +157,32 @@ describe('CaseFlagTableComponent', () => {
     // Check that the second element of the second row of five (i.e. seventh element) does not contain "Decision Reason: "
     expect(tableCellElements.length).toBe(10);
     expect(tableCellElements[6].textContent).not.toContain('Decision Reason: ');
+  });
+
+  it('should display English values for flag name, "Other" description, sub-type value, and comments if in default language', () => {
+    component.flagData = flagData;
+    fixture.detectChanges();
+    const tableCellElements = fixture.debugElement.nativeElement.querySelectorAll('.govuk-table__cell');
+    // Check that the first element of the second row of five (i.e. sixth element) contains content in English
+    expect(tableCellElements.length).toBe(10);
+    expect(tableCellElements[5].textContent).toContain(flagData.flags.details[1].name);
+    expect(tableCellElements[5].textContent).toContain(flagData.flags.details[1].otherDescription);
+    expect(tableCellElements[5].textContent).toContain(flagData.flags.details[1].subTypeValue);
+    // Check that the second element of the second row of five (i.e. seventh element) contains comments in English
+    expect(tableCellElements[6].textContent).toContain(flagData.flags.details[1].flagComment);
+  });
+
+  it('should display Welsh values for flag name, "Other" description, sub-type value, and comments if in Welsh language mode', () => {
+    mockRpxTranslationService.language = 'cy';
+    component.flagData = flagData;
+    fixture.detectChanges();
+    const tableCellElements = fixture.debugElement.nativeElement.querySelectorAll('.govuk-table__cell');
+    // Check that the first element of the second row of five (i.e. sixth element) contains content in Welsh
+    expect(tableCellElements.length).toBe(10);
+    expect(tableCellElements[5].textContent).toContain(flagData.flags.details[1].name_cy);
+    expect(tableCellElements[5].textContent).toContain(flagData.flags.details[1].otherDescription_cy);
+    expect(tableCellElements[5].textContent).toContain(flagData.flags.details[1].subTypeValue_cy);
+    // Check that the second element of the second row of five (i.e. seventh element) contains comments in Welsh
+    expect(tableCellElements[6].textContent).toContain(flagData.flags.details[1].flagComment_cy);
   });
 });