+++ projects/ccd-case-ui-toolkit/src/lib/shared/components/palette/case-flag/components/update-flag/update-flag.component.spec.ts
@@ -110,16 +110,15 @@ describe('UpdateFlagComponent', () => {
     component.formGroup = new FormGroup({
       selectedManageCaseLocation: new FormControl(selectedFlag1)
     });
-
+    nextButton = fixture.debugElement.query(By.css('#updateFlagNextButton')).nativeElement;
     // 200-character text input
     textareaInput = '0000000000' + '1111111111' + '2222222222' + '3333333333' + '4444444444' + '5555555555' + '6666666666' +
       '7777777777' + '8888888888' + '9999999999' + '0000000000' + '1111111111' + '2222222222' + '3333333333' + '4444444444' +
       '5555555555' + '6666666666' + '7777777777' + '8888888888' + '9999999999';
-    // Deliberately omitted fixture.detectChanges() here to allow for a different selected flag to be set for each test
-
-    nextButton = fixture.debugElement.query(By.css('#updateFlagNextButton')).nativeElement;
     // Set default translation language to English
     mockRpxTranslationService.language = 'en';
+    // Deliberately omitted fixture.detectChanges() here to allow for a different selected flag to be set for each test, and
+    // to allow translation language to be set to Welsh for selected tests
   });
 
   it('should create component', () => {
@@ -135,7 +134,7 @@ describe('UpdateFlagComponent', () => {
   });
 
   it('should populate the flag comments textarea from Welsh comments field when only Welsh comments are available', () => {
-    selectedFlag1.flagDetailDisplay.flagDetail.flagComment = null;
+    activeFlag.flagComment = null;
     component.selectedFlag = selectedFlag1;
     fixture.detectChanges();
     // Check the textarea value property, rather than textContent, because this input element has no child nodes
@@ -145,7 +144,7 @@ describe('UpdateFlagComponent', () => {
 
   it('should populate the flag comments textarea from Welsh comments field when selected language is Welsh', () => {
     mockRpxTranslationService.language = 'cy';
-    selectedFlag1.flagDetailDisplay.flagDetail.flagComment = 'First flag';
+    activeFlag.flagComment = 'First flag';
     component.selectedFlag = selectedFlag1;
     fixture.detectChanges();
     // Check the textarea value property, rather than textContent, because this input element has no child nodes
@@ -155,7 +154,7 @@ describe('UpdateFlagComponent', () => {
 
   it('should populate from English comments field when the selected language is Welsh but there are no Welsh comments', () => {
     mockRpxTranslationService.language = 'cy';
-    selectedFlag1.flagDetailDisplay.flagDetail.flagComment_cy = null;
+    activeFlag.flagComment_cy = null;
     component.selectedFlag = selectedFlag1;
     fixture.detectChanges();
     // Check the textarea value property, rather than textContent, because this input element has no child nodes
@@ -164,8 +163,8 @@ describe('UpdateFlagComponent', () => {
   });
 
   it('should show an error message on clicking "Next" if existing comments have been deleted', () => {
-    selectedFlag1.flagDetailDisplay.flagDetail.flagComment = 'First flag';
-    selectedFlag1.flagDetailDisplay.flagDetail.flagComment_cy = null;
+    activeFlag.flagComment = 'First flag';
+    activeFlag.flagComment_cy = null;
     component.selectedFlag = selectedFlag1;
     fixture.detectChanges();
     spyOn(component, 'onNext').and.callThrough();
@@ -192,8 +191,8 @@ describe('UpdateFlagComponent', () => {
   });
 
   it('should show an error message on clicking "Next" if existing comments in Welsh have been deleted', () => {
-    selectedFlag1.flagDetailDisplay.flagDetail.flagComment = null;
-    selectedFlag1.flagDetailDisplay.flagDetail.flagComment_cy = 'Cymraeg';
+    activeFlag.flagComment = null;
+    activeFlag.flagComment_cy = 'Cymraeg';
     component.selectedFlag = selectedFlag1;
     fixture.detectChanges();
     spyOn(component, 'onNext').and.callThrough();