+++ projects/ccd-case-ui-toolkit/src/lib/shared/components/palette/case-flag/pipes/index.ts
@@ -1,1 +1,4 @@
+export * from './flag-field-display.pipe';
+export * from './language-interpreter-display.pipe';
+export * from './manage-case-flags-label-display.pipe';
 export * from './update-flag-title-display.pipe';