+++ RELEASE-NOTES.md
@@ -1,4 +1,7 @@
 ## RELEASE NOTES
+### Version 6.10.7-case-flags-v2-welsh-language-support-additional-rework-2
+**EUI-8116** Additional rework of Welsh language support
+
 ### Version 6.10.7-case-flags-v2-reasonable-adjustments-fixes
 **EUI-8370** Ensure "Other" support type is not shown at the top-level selection when the user is external
 **EUI-8371** Show details of the selected support being updated when the user is external
@@ -12,7 +15,7 @@
 **EUI-8071** Hide "Active flags" banner from external users
 
 ### Version 6.10.7-case-flags-legal-rep-comments-page-v3
-**EUI-7373** Reasonable Adjustments legal rep display and edit ‘Tell us why the support is no longer needed' screen
+**EUI-7373** Reasonable Adjustments legal rep display and edit 'Tell us why the support is no longer needed' screen
 
 ### Version 6.10.7-case-flags-welsh-language-support
 **EUI-7858** Ensure support for Welsh language toggle by storing flag description and comments entries in correct `_cy` fields, and displaying flag comments according to language selection